LayoutTests/platform/ipad/media/modern-media-controls/media-documents/media-document-audio-ios-sizing.html

<script src="../../../../../resources/js-test-pre.js"></script>
<meta name="viewport" content="width=device-width, initial-scale=1">
<body>
<iframe src="../../../../../media/content/silence.mp3" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
<script type="text/javascript">

description("Testing the size of the media element in an audio media document on iOS.");

window.jsTestIsAsync = true;

let media;

(function runTestIfReady() {
    const iframe = document.querySelector("iframe");
    media = iframe.contentDocument.querySelector("video");

    if (!media) {
        setTimeout(runTestIfReady);
        return;
    }

    shouldBecomeEqualToString("getComputedStyle(media).width", "650px", () => {
        shouldBeEqualToString("getComputedStyle(media).height", "50px");

        debug("");
        iframe.remove();
        finishJSTest();
    });

})();

</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
